#37 Refactor the GUI into separate compoments
